adding a bunch of docstrings. replaying headerkeyvaluesanitizer with just a regex version of it.
