+++ tools/test-proxy/Azure.Sdk.Tools.TestProxy/Sanitizers/UriRegexSanitizer.cs
@@ -7,11 +7,19 @@ using System.Threading.Tasks;
 
 namespace Azure.Sdk.Tools.TestProxy.Sanitizers
 {
+    /// <summary>
+    /// General use sanitizer for cleaning URIs via regex.
+    /// </summary>
     public class UriRegexSanitizer : RecordedTestSanitizer
     {
         private Regex _regex;
         private string _newValue;
         
+        /// <summary>
+        /// Runs a regex replace on the member of your choice.
+        /// </summary>
+        /// <param name="regex">The regex used to replace.</param>
+        /// <param name="value">The substituted value.</param>
         public UriRegexSanitizer(string regex, string value = "Sanitized")
         {
             _regex = new Regex(regex, RegexOptions.Compiled);
